+++ src/tests/integration/examples-schema.test.ts
