+++ src/tests/integration/integration-examples.test.ts
@@ -1,6 +1,6 @@
 import test from 'ava';
 import sinon from 'sinon';
-import { processEvents } from '../index';
+import { processEvents } from '../../index';
 import fs from 'fs';
 import path from 'path';
